MySQL: Move binary format fetch outside (#3151)
During a parallel initial load from MySQL to ClickHouse, it was observed
that the PeerDBBinaryFormat call in QValueToAvro which gets the value of
the binary format dynamic setting from catalog was taking ~60% of CPU
utilisation of the flow-worker pod.
Makes sense - making a call to Postgres for every single record can be
expensive
This PR moves this fetch outside the loop on the pulled records.

Also some lint in `cdc_flow.go`