API endpoint for total rows synced by mirror (#3145)
Currently to get the total rows synced by a mirror across CDC and
initial load, clients of our API would have to call
`CDCTableTotalCounts` and `InitialLoadSummary` and in the case of the
latter they would have to sum the `NumRowsSynced` field of all entries.
Not to mention both these API calls would be overfetching if we just
want a single count.

This PR introduces a new endpoint dedicated to getting the sum total
rows synced by CDC and initial load.
An API test has been included.