Do not use slow query from `INFORMATION_SCHEMA`
When the `INFORMATION_SCHEMA` table baloons it will lead to very slow
query execution. In our case - the queries end up taking ~5mins, which
drastically slow down the ingest.

The proposal in this PR is to swap to using a `SHOW COLUMNS IN TABLE
...` and then querying the result from that. There is a Snowflake
documentation link which highlights this problem, and solution as well,
https://community.snowflake.com/s/article/Select-the-list-of-columns-in-the-table-without-using-information-schema

There is an alternative to continue using the `INFORMATION_SCHEMA` but
have a cache for all the tables in the sync, I however, think that will
be trickier to implement.