Another CDC graph endpoint (#3281)
Currently we have a CDC graph where the time period selection denotes
how the data is grouped.
However there is another convention where you'd have the selector
options be "Last hour" or "Last day" etc where the tick size would be 1
minute, 1 hour etc respectively.

This PR introduces a new endpoint in flow api which gives these graph
data points in the above format, and a field for sum of rows for the
given interval of time

---------

Co-authored-by: Philip Dubé <philip.dube@clickhouse.com>