Fetch configuration from DB
No longer passing the entire configuration object to the Temporal job as
we can hit the 2MB hard limit imposed by Temporal. Instead, pass
flowJobName and fetch the config from the DB in all the places required.

There are some additional changes that were required to be made, as some
of the jobs are used from multiple contexts - and I had to adapt some of
the information passed to ensure that they continue to work as expected.