Mongo: limit max length of marshaled floats (#3327)
Numbers like 1e300 (anything > 1e21) technically pass the check of being
an integer but scientific notation would be a better hint than spelling
out all their digits with a `.0`, also avoiding the size of s3 payload
and the raw table getting ballooned.

We're not expecting a lot of such values but just something for peace of
mind.