Better flow statuses for renaming and table removal phases  (#3343)
## Why
There are currently two states in the lifecycle of a mirror where the
status is `Running`/`Snapshot` but the mirror is not in those states.

### Table removal

Currently, when a mirror is in the process of removing tables, the
status shows up as `Running`, when in reality the mirror is not in a
position to accept flow state change signals; if you were to request a
table removal during an ongoing table removal, your request would
"succeed" as far as the API call goes but the signal is not acted upon
by the worker.

A consequence and another pain point is that there is no way to
programmatically know if a mirror is finished with its table removal as
the state is always Running.

### Renaming phase of resync

The last step of a resync is an activity where we exchange the _resync
tables with the originals. Much like table removal, this is a state
where the mirror cannot be paused/edited.

## What
- This PR introduces a `MODIFYING` flow state for table removal.
- Sets the status to `RESYNC` during renaming.

- [ ] Automated and functional testing pending