CDC-Only table addition support exposed via Flow API (#3418)
# Why
- Currently, table additions in mirrors without initial load are having
their initial loads be done. We have received feedback that this is
unideal and comes across as a bug. This lays the foundation for exposing
an opt-in/out flag in PeerDB client software for initial load for added
tables.
- In multi-tenant source data store setups, when migrating from one
PeerDB instance to another, it is often desirable to migrate one tenant
at a time, using the same replication slot.

# What
- Adds a field to the FlowConfigUpdate object (which is the request for
table editing) to skip initial load for table additions. This way this
PR is backwards-compatible versus `doInitialSnapshotForAddedTables`.
- `doInitialSnapshot` will be set to the opposite of above value for
table addition workflow
- Adds an API test for this feature.

I plan to expose this via PeerDB UI as a follow-up