fix(ui): optimize SchemaBox avoid fetch table schema when collapsed
SchemaBox component is used for table mapping in mirror creation flow. When component is mounted, all the schemas are collapsed. However, what is currently observed, regardless of box is collapsed, `fetchTablesForSchema` is always executed causing overhead requests being sent to API.

Desired behaviour would be tables are fetched only if box is being expanded. This PR fixes it, trying to keep the functionality of reloading results when the `initialLoadOnly` prop is changed. It has a side effect of retried request if we collapse and expand box again, but I think it's an edge case.

It's not a pain in most cases (it's less likely to have that many schemas it overloads API), but for my testing I had a peer source that consisted of >100 schemas.