Move shared validations into their own module (#3547)
This allows to import just the validation code with minimal dependencies
and avoid tracking unrelated dependency updates in PeerDB.

Connector-specific code remaining in shared/ was kind of small so moved
it to other places. CH type conversion was supposed to be shared with
other services in the future but it pulls in QValue so we can decide to
move all of it to shared-light if we want to expose QValue or to move
just type conversions and give it its own small data contracts.