Proposal: Activity to fetch config from the DB
This uses a new helper in `internal.FetchConfigFromDB` to fetch a fully
hydrated `protos.FlowConnectionConfigs`. When passing this config to
Temporal we strip the `tableMappings` array element which can cause it
to go over the 2MB limit.

This contains a subset of the changes which were originally proposed in:
https://github.com/PeerDB-io/peerdb/pull/3407

Tenanative first go at removing `tableMappings` & storing in DB