Decouple FlowConnectionConfigs from API (#3605)
In #3589 we want to stop using a field in a way that's not error-prone
in the rest of the code (so, removing) but also doesn't break API
backcompat (so, not removing).

Trying to split FlowConnectionConfigs into two contracts that are made
sure to be in sync with a little codegen. Generated code has small TODOs
that would be removed by #3589 if we go this route.

Validated that renaming the type of a workflow arg works well with
history replay, only the payload is important.