make settings versioning more robust (#3614)
- Make sure `type_json_skip_duplicated_paths` is also conditionally set
only if clickhouse version supports it.
- Create a setting_util.go file to centralize settings definition and
minimum versioning, to make setting generation more robust. This has
been an issue in the past, and we'll likely need to continue to add more
settings later.
- Separately, added some comments for how PeerDB Internal Version System
works


We also have a number of settings that we set at the client level
[here](https://github.com/PeerDB-io/peerdb/blob/a6d0527bb3443dac36097e0a47affc286c8eab23/flow/connectors/clickhouse/clickhouse.go#L243-L260).
Not sure if client automatically set settings based on server version,
but since we don't check ch version there explicitly in code, that seems
to be the case. The caveats of these settings is that they are not
backwards compatible, since this is on the peer creation flow and not
mirror creation flow, so internal version is not currently available if
needed.

The change in this PR only affects query-level settings. It may be
possible that some of these query-level SETTINGs could instead be
client-level (i.e. `type_json_skip_duplicated_paths`), but query-level
settings seem to be more flexible with backwards compatibility support.