Set wal_sender_timeout=120s on CDC startup (#3682)
We've determined that this setting is safe to set and will reduce the
impact of random disconnects (see Nov 11 clickpipes knowledge share).

Defaulting to 120s, with a config to override or not set, that we can
use if something breaks.

Validated that it works as intended from a dev PeerDB to RDS - network
cut, reconnect, RDS lets us in, and with the server default timeout 0 it
doesn't for 15-20 minutes.