fix(bigquery): fix down-scoped token prefix path
When a root path of bucket is provided `gs://my_bucket/` as a staging bucket, a GCP down scoped token contains wrong bucket filter expression.

See `storageDownScopedToken`:

https://github.com/PeerDB-io/peerdb/blob/a6da92fc3839342e7e0cc67794d05e97fa6c2131/flow/connectors/bigquery/gcp.go#L15-L36

The reason is when path is empty, condition is produced like:

`resource.name.startsWith('projects/_/buckets/my_bucket/objects//')`

but it should be:

`resource.name.startsWith('projects/_/buckets/my_bucket/objects/')`

It's an easy fix and comes with a test change to ensure we cover that case in future.