Improve initial load partitioning logic readability in PG (#3687)
We have expanded our partitioning logic over time to the point that
getNumRowsPartitions has gotten quite long, so seems like a good time to
break it into more digestible pieces.

I also noticed we were only committing the transaction for the default
path, but not for the alternative path with numPartitionsOverride, which
doesn't have an impact on correctness because of defer
shared.RollbackTx(), but was confusing nonetheless. Moved the "commit"
logic to GetQRepPartitions.

The refactor introduces a light function interface to implement
partitioning logic. This will make testing/exposing different strategies
easier later on. (i.e. if/when we do plan to expose CTID partitioning).

Separately, I think we can use estimated table count instead of SELECT
count(*) for NTileBucketPartitioning, which @serprex proposed before.
The implementation doesn't need a precise row count. So this should help
with improving performance of this count query. Leaving that for a
follow-up PR.

Test:
- local smoke testing to ensure it works as expected