Feature Flag for skipping Avro compression algorithm
We are observing a return of the Snowflake errors:
`COPY INTO command: 100079 (22000): Invalid data encountered during
decompression for file: 'not_file',compression type used: 'ZSTD', cause:
'Data corruption detected'`

These were previously resolved by using `AUTO_COMPRESS=false`; however
it seems that in the meantime further changes make the `zst` compression
applied to Avro files to no longer be in compliance with Snowflake,
leading to failures to decode the files.

This change introduces a feature flag that allows for the skipping of
the `zstd` compression on Avro files, making the files that land in
Snowflake fully uncompressed.