[snowflake] Feature Flag for skipping Avro compression algorithm (#3707)
We are observing a return of the Snowflake errors: `COPY INTO command:
100079 (22000): Invalid data encountered during decompression for file:
'not_file',compression type used: 'ZSTD', cause: 'Data corruption
detected'`

These were previously resolved by using `AUTO_COMPRESS=false`; however
it seems that in the meantime further changes make the `zst` compression
applied to Avro files to no longer be in compliance with Snowflake,
leading to failures to decode the files.

This change introduces a feature flag that allows for the skipping of
the `zstd` compression on Avro files, making the files that land in
Snowflake fully uncompressed.

---

**NB:**
We are successfully running v0.30.9 in production w/o this error
happening, however on updating to v0.34.8 the error returns. I have
attempted to find out where between v0.30.9 and v0.34.8 there was the
breaking change which led to this issue returning however I was not
successful and ran out of allocated time - so I'm suggesting this
alternative which would allow us to move to a more recent version of
`peerdb`.