track uncompressed bytes for mongodb connector (#3715)
MongoDB Avro file has a very high compression ratio. This can result in
OOM issues during file upload to ClickHouse during initial snapshot. The
proposed solution compute the uncompressed bytes of QRecords based on
Avro encoding. It reuses the `PEERDB_S3_BYTES_PER_AVRO_FILE` variable to
determine limit, and then based on connector source type, determines
whether to compute the compressed bytes (which is conveniently computed
by the WatchWriter and how we've tracked this historically) or
uncompressed bytes (in the case for MongoDB Connector, where we manually
compute bytes as we process QValues). If we want a different default, we
could also introduce a new env variable instead, e.g.
`PEERDB_S3_UNCOMPRESSED_BYTES_PER_AVRO_FILE`. Keeping it to initial
snapshot for now since that's where OOM issue is happening. If we want
to extend this to CDC in the future, the computation would also be
different due to the schema of the staging table.

Add unit tests
Run manual smoke test