Cancel table addition feature (#3735)
## Overview
This PR introduces a new feature -- cancelling ongoing table additions
of CDC mirrors. Specifically, if a mirror has been paused and is adding
tables but is stuck in the `Setup` or `Snapshot` phases, it can be
brought back to CDC aka `Running`, and any tables as part of the
addition request which finished snapshot will be part of the restored
CDC.

## Why
Suppose you were to pause your CDC mirror and add 5 tables. Table
addition of those 5 tables involve setup of those 5 tables where they
are created on target, and then snapshot aka initial load. The
underlying mirror is paused during this process to maintain data
consistency. This can be problematic for a few reasons:
- Your destination tables there are growing stale.
- In the case of Postgres sources, your replication slot is growing and
eating up your storage. Also, if a network interruption was to occur,
the snapshot is invalid and you have no choice but to resync/create a
new mirror.
- In the case of MySQL and MongoDB sources, your replication logs are
getting closer to expiry.

We have seen cases where users add very large tables which take time and
the above issues are more pronounced, making them wish they could cancel
the table addition and have CDC heal their pipeline.

## What

Exposed via an API endpoint, this PR introduces a new
CancelTableAdditionFlow temporal workflow. At a high-level, the workflow
steps are as follows:
- Takes in as input the mirror name and a list of currently replicating
tables.
- Checks the user-facing `qrep_runs` table to get a list of tables which
have completed snapshot for this mirror.
- Cleans up the incomplete tables in `qrep_runs` 
- Obtains source table OIDs from catalog, owing to
https://github.com/PeerDB-io/peerdb/pull/3704
- Uses the information obtained above to construct the required input to
a new CDCFlowWorkflow with the same name as the originally running CDC
flow. We construct the input in. a way where this new CDC flow skips
setup and snapshot and directly goes to `Running`.

An end-to-end test has been added, and more are pending to be written.

---------

Co-authored-by: Ilia Demianenko <ilia.demianenko@clickhouse.com>