MySQL: support gh-ost migrations (#3756)
[gh-ost](https://github.com/github/gh-ost) is a popular MySQL schema
migration tool. It does migrations for a table `A` via creating a new
table `A_gho`, syncing everything into it via binlog (just like us),
then doing an [atomic
rename](https://github.com/github/gh-ost/issues/82) `A to A_old, A_gho
to A`. For us it means exactly nothing table identifier-wise as we use
names, but we'll miss the column adds that would come out of it.

This change replaces silent drop of unknown columns with constructing a
table schema delta from the TABLE_MAP_EVENT that we get with every
ROWS_EVENT.