Test mysql 9.5 (gtid) and 5.7 (filepos)  (#3767)
Testing mysql 5.7 uncovered a number of scenarios that are not properly
supported by (1) older version of mysql (2) when binlog_row_metadata is
set to MINIMAL.

Lacking `binlog_row_metadata` means no proper support for:
- Remove column
- Ghost schema changes
- Column exclusion
- Enum

Other things learned:
- mysql5 does not support utf8mb4 in JSON, so emoji doesn't show up
properly
- only mysql9+ support Vector data type

Conclusion is that supporting `MINIMAL` opens up a can of worms in terms
of edge cases. While we support mysql 5 due to its popularity, it is EOL
since 2 years ago, and it's not a great idea to allow 'MINIMAL' for
mysql 8+.


In the process, did small refactoring of mysql setup.