fix(bigquery): improve RECORD data type handling for BigQuery connector (#3774)
Avro-record types converted to Tuple fails to be ingested into
ClickHouse into a String column. From now, all RECORD fields will be
force-normalized with CAST. RECORD array will be normalized into
Array<String>. This requires `QField` to be extended with extra
information about original data type. This will be optional and used for
BQ only.

Another change is for Avro row number estimation for partitioning.
Current implementation reads data and decodes it. This is fragile and
out of control. It's replaced with a no data read header decoding for
row numbers.

As a bonus, this PR fixes schema inference for PeerDB UI. (more
information about BQ fields in schema information)