put toast tracking behind a FF (#3790)
Make it possible to skip TOAST tracking for Postgres. Currently the
TOAST backfill implementation is incomplete as it does not guarantee
TOAST column replication anyways; and it can introduce memory/disk
pressure when TOAST data is large. The FF makes it possible to bypass it
as needed. Proper TOAST implementation without replica identity full
will need to be redesigned.

This PR is based on https://github.com/PeerDB-io/peerdb/pull/3388 .
Leaving the default as enabled for now, will evaluate a rollout plan for
disabling CDCStore to not introduce any regression.

---------

Co-authored-by: Ilia Demianenko <ilia.demianenko@clickhouse.com>